Add runtime build status balls to External Console Log actions.
Also generalize the solution
