+++ src/main/resources/jenkins/plugins/logstash/kibana/AbstractConsoleAction/buildCaption.jelly
@@ -1,0 +1,53 @@
+<!--
+The MIT License
+
+Copyright (c) 2004-2009, Sun Microsystems, Inc., Kohsuke Kawaguchi
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+-->
+
+<!--
+    displays a caption for build/externalRun.
+-->
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+    <h1 class="build-caption page-headline">
+      <j:if test="${it.run.building}">
+        <div class="build-caption-progress-container" style="float:right">
+          <table class="middle-align"><tr>
+            <td>
+              ${%Progress}:
+            </td><td class="build-caption-progress-bar">
+              <t:buildProgressBar build="${it.run}"/>
+            </td><td>
+              <!--
+              <j:if test="${it.run.parent.hasAbortPermission()}">
+                <l:stopButton href="stop" confirm="${%confirm(it.run.fullDisplayName)}" alt="[${%cancel}]"/>
+              </j:if>
+              -->
+            </td>
+          </tr></table>
+        </div>
+      </j:if>
+
+      <l:icon alt="${it.run.iconColor.description}" class="${it.run.buildStatusIconClassName} icon-xlg" tooltip="${it.run.iconColor.description}"/>
+      ${%Console Output from} ${it.dataSourceDisplayName}
+    </h1>
+</j:jelly>
+
